requirements.txt
numpy
tqdm
einops
tensorboardX
matplotlib
pyyaml
h5py
